Create seperate files for lexer and start
